# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.QuranKhatamCycle:
    properties:
      completed_at:
        example: "2026-02-01T00:00:00Z"
        type: string
      completed_juz:
        items:
          type: integer
        type: array
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      juz_count:
        example: 17
        type: integer
      notes:
        example: Khatam Ramadhan
        type: string
      percent:
        example: 56.67
        type: number
      started_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  entity.Translation:
    properties:
      destination:
        example: en
        type: string
      original:
        example: текст для перевода
        type: string
      source:
        example: auto
        type: string
      translation:
        example: text for translation
        type: string
    type: object
...
  entity.TranslationHistory:
    properties:
      history:
        items:
          $ref: '#/definitions/entity.Translation'
        type: array
    type: object
...
  v1.Error:
    properties:
      code:
        example: invalid_request_body
        type: string
      details: {}
      error:
        example: message
        type: string
      message:
        example: message
        type: string
      request_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      retry_after:
        example: 60
        type: integer
    type: object
...
  v1.KhatamHistory:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.QuranKhatamCycle'
        type: array
      total:
        example: 2
        type: integer
    type: object
...
  v1.StartKhatamCycle:
    properties:
      notes:
        example: Khatam Ramadhan
        maxLength: 2000
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Surau classical book reader API
  title: Go Clean Template API
  version: "1.0"
paths:
...
  /me/quran/khatam:
    get:
      description: Return the authenticated user's active khatam cycle with completed
        juz marks.
      operationId: get-active-khatam-cycle
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.QuranKhatamCycle'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Get active khatam cycle
      tags:
      - me
    post:
      consumes:
      - application/json
      description: Start a new Quran khatam cycle. Only one active cycle is allowed
        per user.
      operationId: start-khatam-cycle
      parameters:
      - description: Optional cycle notes
        in: body
        name: request
        schema:
          $ref: '#/definitions/v1.StartKhatamCycle'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.QuranKhatamCycle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Start khatam cycle
      tags:
      - me
  /me/quran/khatam/complete:
    post:
      description: Complete the active khatam cycle. Requires all 30 juz to be marked;
        marking the 30th juz does not auto-complete so accidental marks stay reversible.
      operationId: complete-khatam-cycle
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.QuranKhatamCycle'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Complete khatam cycle
      tags:
      - me
  /me/quran/khatam/history:
    get:
      description: Return the authenticated user's completed khatam cycles ordered
        by completion recency.
      operationId: list-khatam-history
      parameters:
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.KhatamHistory'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: List khatam history
      tags:
      - me
  /me/quran/khatam/juz/{juz_number}:
    delete:
      description: Remove one juz mark (1-30) from the active khatam cycle. Idempotent.
      operationId: unmark-khatam-juz
      parameters:
      - description: Juz number (1-30)
        in: path
        name: juz_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.QuranKhatamCycle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Unmark khatam juz
      tags:
      - me
    put:
      description: Mark one juz (1-30) as completed on the active khatam cycle. Idempotent.
      operationId: mark-khatam-juz
      parameters:
      - description: Juz number (1-30)
        in: path
        name: juz_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.QuranKhatamCycle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Mark khatam juz
      tags:
      - me
...
  /translation/history:
    get:
      description: Show all translation history for current user
      operationId: history
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.TranslationHistory'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Show history
      tags:
      - translation
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"